Added PID powered paddle steering.
Finally, we had to implement the PID strategy,
with this commit, I've added a PID controller class, and PID balancer class.
Which are used to implement the desired
way of steering the paddle.
In order to make the solution cleaner, some new
methods were added to the Paddle class.
Also, the default ball position was fixed.