Derivative calculation method changed.
Previous way of calculating derivative in PID,
was not correct. Changes in the error vector
were too small, and it often resulted with the derivative
error value being 0. In this commit I've come up with
a temporary solution to calculate the average value of derivative.
It helps to smooth the calculated values, but definitely need
more attention and observation.